ARROW-10325: [C++][Compute] Refine aggregate kernel registration
Separate Mode and Variance/Stddev kernels registration from basic
aggregation kernels.
