+++ cpp/src/arrow/compute/registry.cc
@@ -128,6 +128,8 @@ static std::unique_ptr<FunctionRegistry> CreateBuiltInRegistry() {
 
   // Aggregate functions
   RegisterScalarAggregateBasic(registry.get());
+  RegisterScalarAggregateMode(registry.get());
+  RegisterScalarAggregateVariance(registry.get());
 
   // Vector functions
   RegisterVectorHash(registry.get());
