+++ cpp/src/arrow/compute/registry_internal.h
@@ -43,6 +43,8 @@ void RegisterVectorSort(FunctionRegistry* registry);
 
 // Aggregate functions
 void RegisterScalarAggregateBasic(FunctionRegistry* registry);
+void RegisterScalarAggregateMode(FunctionRegistry* registry);
+void RegisterScalarAggregateVariance(FunctionRegistry* registry);
 
 }  // namespace internal
 }  // namespace compute
